feat: 🎸 add density to header
