+++ src/Molecules/Table/Header/Header.tsx
@@ -1,7 +1,28 @@
 import React from 'react';
-import { Props } from './Header.types';
+import styled from 'styled-components';
+import { Props, Density } from './Header.types';
 import { Flexbox, Typography } from '../../../index';
 import { isElement } from '../../../common/utils';
+
+const getDensityPaddings = (density: Density) => {
+  switch (density) {
+    case 's':
+      return 0;
+    case 'm':
+      return 1;
+    case 'l':
+      return 2;
+    default:
+      return 1;
+  }
+};
+
+const StyledFlexbox = styled(Flexbox).withConfig({
+  shouldForwardProp: prop => !['density'].includes(prop),
+})<{ density: Density }>`
+  padding-top: ${p => p.theme.spacing.unit(getDensityPaddings(p.density))}px;
+  padding-bottom: ${p => p.theme.spacing.unit(getDensityPaddings(p.density))}px;
+`;
 
 export const Header: React.FC<Props> = ({
   className,
@@ -13,11 +34,12 @@ export const Header: React.FC<Props> = ({
   shrink,
   wrap = 'nowrap',
   fontSize = 's',
+  density = 'm',
   children,
   ...htmlProps
 }) => {
   return (
-    <Flexbox
+    <StyledFlexbox
       container={container}
       flex={flex}
       grow={grow}
@@ -26,6 +48,7 @@ export const Header: React.FC<Props> = ({
       shrink={shrink}
       wrap={wrap}
       className={className}
+      density={density}
       role="columnheader"
       {...htmlProps}
     >
@@ -36,6 +59,6 @@ export const Header: React.FC<Props> = ({
           {children}
         </Typography>
       )}
-    </Flexbox>
+    </StyledFlexbox>
   );
 };
