+++ src/Molecules/Table/Row/Row.types.ts
@@ -2,10 +2,7 @@ import React, { ReactNode } from 'react';
 
 type HtmlProps = {} & React.HTMLAttributes<HTMLDivElement>;
 
-export type Density = 's' | 'm' | 'l';
-
 type Props = {
-  density?: Density;
   hideBorder?: boolean;
   expandableContent?: ReactNode;
   expanded?: boolean;
